1. onRollback check load flag first

+++ tinker-server-sample/src/main/java/tinker/sample/android/patchserver/SamplePatchRequestCallback.java
@@ -104,7 +104,11 @@ public class SamplePatchRequestCallback extends DefaultPatchRequestCallback {
     public void onPatchRollback() {
         TinkerLog.w(TAG, "onPatchRollback");
         TinkerServerClient client = TinkerServerClient.get();
-
+        Tinker tinker = client.getTinker();
+        if (!tinker.isTinkerLoaded()) {
+            TinkerLog.w(TAG, "onPatchRollback, tinker is not loaded, just return");
+            return;
+        }
         if (Utils.isBackground()) {
             TinkerLog.i(TAG, "onPatchRollback, it is in background, just clean patch and kill all process");
             rollbackPatchDirectly();
